reorganize join code
